Move the function related to rendering the header to the Header Renderer module
